requirements.txt
selenium==3.141.0
beautifulsoup4==4.9.3
pendulum==2.1.2
pyyaml==5.4.1
pyvirtualdisplay==0.2.5
dynaconf==3.1.2